Change session summary writer function

+++ model.py
@@ -174,17 +174,17 @@ class CycleGAN(object):
 
     def summary(self):
         with tf.name_scope('generator_summaries'):
-            cycle_loss_summary = tf.summary.scalar('cycle_loss', self.cycle_loss)
-            identity_loss_summary = tf.summary.scalar('identity_loss', self.identity_loss)
-            generator_loss_A2B_summary = tf.summary.scalar('generator_loss_A2B', self.generator_loss_A2B)
-            generator_loss_B2A_summary = tf.summary.scalar('generator_loss_B2A', self.generator_loss_B2A)
-            generator_loss_summary = tf.summary.scalar('generator_loss', self.generator_loss)
+            cycle_loss_summary = tf.summary.scalar('cycle_loss',tf.cast(self.cycle_loss,tf.float32))
+            identity_loss_summary = tf.summary.scalar('identity_loss', tf.cast(self.identity_loss,tf.float32))
+            generator_loss_A2B_summary = tf.summary.scalar('generator_loss_A2B', tf.cast(self.generator_loss_A2B,tf.float32))
+            generator_loss_B2A_summary = tf.summary.scalar('generator_loss_B2A', tf.cast(self.generator_loss_B2A,tf.float32))
+            generator_loss_summary = tf.summary.scalar('generator_loss', tf.cast(self.generator_loss,tf.float32))
             generator_summaries = v1.summary.merge([cycle_loss_summary, identity_loss_summary, generator_loss_A2B_summary, generator_loss_B2A_summary, generator_loss_summary])
 
         with tf.name_scope('discriminator_summaries'):
-            discriminator_loss_A_summary = tf.summary.scalar('discriminator_loss_A', self.discriminator_loss_A)
-            discriminator_loss_B_summary = tf.summary.scalar('discriminator_loss_B', self.discriminator_loss_B)
-            discriminator_loss_summary = tf.summary.scalar('discriminator_loss', self.discriminator_loss)
+            discriminator_loss_A_summary = tf.summary.scalar('discriminator_loss_A', tf.cast(self.discriminator_loss_A,tf.float32))
+            discriminator_loss_B_summary = tf.summary.scalar('discriminator_loss_B', tf.cast(self.discriminator_loss_B,tf.float32))
+            discriminator_loss_summary = tf.summary.scalar('discriminator_loss', tf.cast(self.discriminator_loss,tf.float32))
             discriminator_summaries = v1.summary.merge([discriminator_loss_A_summary, discriminator_loss_B_summary, discriminator_loss_summary])
 
         return generator_summaries, discriminator_summaries
